add algo fun. section
